requirements.txt
discord.py==1.7.3
pandas==1.3.4
python-dotenv==0.19.2
requests==2.26.0
